+++ src/Cuba-4.2-nonwin/src/common/Fork.c
@@ -20,7 +20,9 @@
 
 #define MINCORES 1
 
-coreinit cubafun_;
+/* Commented out for gcc 10.0 -fno-common default */
+/* coreinit cubafun_; */
+
 extern int cubaverb_;
 extern corespec cubaworkers_;